itins/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
celery==5.3.6
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-modeltranslation==0.18.11
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
GDAL==3.4.1
pillow==10.2.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
pymemcache==4.0.0
requests==2.31.0
selenium==4.16.0
sqlparse==0.4.4
timedelta==2020.12.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
zope.interface==6.1